// src/types/base/error-code.type.ts
export enum ErrorCode {
  SYSTEM_ERROR = 'IEM00', // Đã xảy ra lỗi, vui lòng thử lại sau
  INVALID_CREDENTIALS = 'IEM01', // Thông tin đăng nhập không hợp lệ
  USER_NOT_FOUND = 'IEM02', // User không tồn tại
  AUTH_FAILED = 'IEM03', // Lỗi xác thực
  CUSTOMER_CREATE_FAILED = 'IEM04', // Lỗi tạo khách hàng
  CUSTOMER_NOT_FOUND = 'IEM05', // Khách hàng không tồn tại
  CUSTOMER_ADDRESS_CREATE_FAILED = 'IEM06', // Lỗi tạo địa chỉ khách hàng
  CUSTOMER_CONTACT_CREATE_FAILED = 'IEM07', // Lỗi tạo liên hệ khách hàng
  CUSTOMER_CONTACT_DELETE_FAILED = 'IEM08', // Lỗi xóa liên hệ khách hàng
  CUSTOMER_CONTACT_UPDATE_FAILED = 'IEM09', // Lỗi cập nhật liên hệ khách hàng
  CUSTOMER_CONTACT_NOT_FOUND = 'IEM10', // Liên hệ khách hàng không tồn tại
  CUSTOMER_ADDRESS_NOT_FOUND = 'IEM11', // Địa chỉ khách hàng không tồn tại
  CUSTOMER_ADDRESS_DELETE_FAILED = 'IEM12', // Lỗi xóa địa chỉ khách hàng
  CUSTOMER_ADDRESS_UPDATE_FAILED = 'IEM13', // Lỗi cập nhật địa chỉ khách hàng
  CUSTOMER_SEARCH_FAILED = 'IEM14', // Lỗi tìm kiếm khách hàng
  CREATE_FAILED = 'IEM15', // Lỗi tạo
  UPDATE_FAILED = 'IEM16', // Lỗi cập nhật
  NOT_FOUND = 'IEM17', // Không tồn tại
  DELETE_FAILED = 'IEM18', // Lỗi xóa
  SEARCH_FAILED = 'IEM19', // Lỗi tìm kiếm
  GET_ALL_FAILED = 'IEM20', // Lỗi lấy tất cả
  USER_EXISTS = 'IEM21', // Tên đăng nhập đã tồn tại
  FORBIDDEN_MESSAGE = 'IEM22', // Không có quyền truy cập
  UPSERT_ERROR = 'IEM23', // Lỗi upsert
  MISSING_DEVICE_STATUS = 'IEM24', // Thiếu trạng thái thiết bị
  INVALID_ARGUMENT = 'IEM25', // Lỗi đối số không hợp lệ
  DEVICE_NOT_IN_STOCK = 'IEM26', // Thiết bị không còn trong kho
  DEVICE_ALREADY_IN_STOCK = 'IEM27', // Thiết bị đã có trong kho
  FIND_STOCK_ERROR = 'IEM28', // Lỗi tìm kiếm tồn kho
  DEVICE_NOT_IN_WAREHOUSE = 'IEM29', // Thiết bị không còn trong kho
  INVALID_DATA = 'IEM30', // Dữ liệu không hợp lệ
  INVALID_STATUS = 'IEM31', // Trạng thái không hợp lệ
  FIND_BY_ID_FAILED = 'IEM32', // Lỗi tìm kiếm theo id
  FIND_BY_CODE_FAILED = 'IEM33', // Lỗi tìm kiếm theo code
  TRANSACTION_RECEIPT_CONFIRMED = 'IEM34', // Phiếu nhập kho đã được xác nhận
  TRANSACTION_RECEIPT_ISSUE_CONFIRMED = 'IEM35', // Phiếu xuất kho đã được xác nhận
  INVALID_ORDER_STATUS = 'IEM36', // Trạng thái đơn hàng không hợp lệ
  PRICE_LIST_NOT_FOUND = 'IEM37', // Giá cả không tồn tại
  DEVICE_NOT_FOUND = 'IEM38', // Thiết bị không tồn tại
  INVOICE_NOT_FOUND = 'IEM39', // Hóa đơn không tồn tại
  DEVICE_ALREADY_SUBSCRIBED = 'IEM40', // Thiết bị đã được đăng ký
  PRODUCT_OUT_OF_STOCK = 'IEM41', // Sản phẩm hết hàng
  DEVICE_NOT_FOUND_IN_RETURN_ORDER = 'IEM42', // Thiết bị không tồn tại trong đơn hàng trả hàng
  SUBSCRIPTION_DEVICE_NOT_FOUND = 'IEM43', // Thiết bị không tồn tại trong đơn hàng đăng ký
  SUBSCRIPTION_NOT_FOUND = 'IEM44', // Đơn hàng đăng ký không tồn tại
  NOT_ENOUGH_STOCK = 'IEM45', // Sản phẩm không đủ hàng
  SUBSCRIPTION_NOT_ACTIVE = 'IEM46', // Đơn hàng đăng ký không hoạt động
  DEVICE_NOT_AVAILABLE = 'IEM47', // Thiết bị không còn khả dụng
}